Extract row builder from repo.ts
